// models/AggComposite.ts
/**
 * Manticore Search Client
 * Сlient for Manticore Search. 
 *
 * OpenAPI spec version: 5.0.0
 * Contact: info@manticoresearch.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { AggCompositeSource } from '../models/AggCompositeSource';
import { HttpFile } from '../http/http';

/**
* Object to perform composite aggregation, i.e., grouping search results by multiple fields
*/
export class AggComposite {
    /**
    * Maximum number of composite buckets in the result
    */
    'size'?: number;
    'sources'?: Array<{ [key: string]: AggCompositeSource; }>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "size",
            "baseName": "size",
            "type": "number",
            "format": ""
        },
        {
            "name": "sources",
            "baseName": "sources",
            "type": "Array<{ [key: string]: AggCompositeSource; }>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AggComposite.attributeTypeMap;
    }

    public constructor() {
    }
}