// models/AggCompositeSource.ts
/**
 * Manticore Search Client
 * Сlient for Manticore Search. 
 *
 * OpenAPI spec version: 5.0.0
 * Contact: info@manticoresearch.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { AggCompositeTerm } from '../models/AggCompositeTerm';
import { HttpFile } from '../http/http';

/**
* Object containing terms used for composite aggregation.
*/
export class AggCompositeSource {
    'terms': AggCompositeTerm;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "terms",
            "baseName": "terms",
            "type": "AggCompositeTerm",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AggCompositeSource.attributeTypeMap;
    }

    public constructor() {
    }
}