// models/AggCompositeTerm.ts
/**
 * Manticore Search Client
 * Сlient for Manticore Search. 
 *
 * OpenAPI spec version: 5.0.0
 * Contact: info@manticoresearch.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Object representing a term to be used in composite aggregation.
*/
export class AggCompositeTerm {
    /**
    * Name of field to operate with
    */
    'field': string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "field",
            "baseName": "field",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AggCompositeTerm.attributeTypeMap;
    }

    public constructor() {
    }
}