// models/AggHistogram.ts
/**
 * Manticore Search Client
 * Сlient for Manticore Search. 
 *
 * OpenAPI spec version: 5.0.0
 * Contact: info@manticoresearch.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

/**
* Object to use histograms in aggregation, i.e., grouping search results by histogram values
*/
export class AggHistogram {
    /**
    * Field to group by
    */
    'field': string;
    /**
    * Interval of the histogram values
    */
    'interval': number;
    /**
    * Offset of the histogram values. Default value is 0.
    */
    'offset'?: number;
    /**
    * Flag that defines if a search response will be a dictionary with the bucket keys. Default value is false.
    */
    'keyed'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "field",
            "baseName": "field",
            "type": "string",
            "format": ""
        },
        {
            "name": "interval",
            "baseName": "interval",
            "type": "number",
            "format": ""
        },
        {
            "name": "offset",
            "baseName": "offset",
            "type": "number",
            "format": ""
        },
        {
            "name": "keyed",
            "baseName": "keyed",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return AggHistogram.attributeTypeMap;
    }

    public constructor() {
    }
}