// models/BoolFilter.ts
/**
 * Manticore Search Client
 * Сlient for Manticore Search. 
 *
 * OpenAPI spec version: 5.0.0
 * Contact: info@manticoresearch.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { QueryFilter } from '../models/QueryFilter';
import { HttpFile } from '../http/http';

export class BoolFilter {
    /**
    * Query clauses that must match for the document to be included
    */
    'must'?: Array<QueryFilter>;
    /**
    * Query clauses that must not match for the document to be included
    */
    'must_not'?: Array<QueryFilter>;
    /**
    * Query clauses that should be matched, but are not required
    */
    'should'?: Array<QueryFilter>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "must",
            "baseName": "must",
            "type": "Array<QueryFilter>",
            "format": ""
        },
        {
            "name": "must_not",
            "baseName": "must_not",
            "type": "Array<QueryFilter>",
            "format": ""
        },
        {
            "name": "should",
            "baseName": "should",
            "type": "Array<QueryFilter>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BoolFilter.attributeTypeMap;
    }

    public constructor() {
    }
}